Refactor: Move script injection logic to jsCookie and update PrebidInterface for improved structure
