+++ packages/extension/src/contentScript/index.ts
@@ -102,28 +102,6 @@ class WebpageContentScript {
     this.docElement = document.documentElement;
 
     this.listenToConnection();
-
-    /**
-     * When the content script for the cookies is loaded the web_accessible resource is injected
-     * in the page by the content script. The script is then run and when everything from the script
-     * is loaded and completed. The script is removed from the page. This helps in avoiding tampering with the script.
-     */
-    const injectScript = () => {
-      const script = document.createElement('script');
-      script.src = chrome.runtime.getURL('/prebid-interface.js');
-      const node = document.head || document.documentElement;
-
-      if (node) {
-        node.appendChild(script);
-        script.onload = () => {
-          script.remove();
-        };
-      } else {
-        requestIdleCallback(injectScript);
-      }
-    };
-
-    injectScript();
   }
 
   /**
