+++ packages/extension/src/manifest.json
@@ -41,7 +41,8 @@
     {
       "js": ["js-cookie-content-script.js"],
       "matches": ["*://*/*"],
-      "all_frames": true
+      "all_frames": true,
+      "run_at": "document_start"
     }
   ],
   "action": {
